+++ src/controllers/GameHistory.js
@@ -83,3 +83,5 @@ exports.getLastProtectedPlayer = async gameId => {
 exports.getLastVotePlay = gameId => this.findOne({ gameId, "play.action": "vote" }, null, { sort: { createdAt: -1 } });
 
 exports.getLastSistersPlay = gameId => this.findOne({ gameId, "play.source": "two-sisters" }, null, { sort: { createdAt: -1 } });
+
+exports.getLastBrothersPlay = gameId => this.findOne({ gameId, "play.source": "three-brothers" }, null, { sort: { createdAt: -1 } });