+++ tests/e2e/game/full-game.test.js
@@ -23,6 +23,9 @@ let players = [
     { name: "Dêg", role: "cupid" },
     { name: "Dæg", role: "two-sisters" },
     { name: "D∂g", role: "two-sisters" },
+    { name: "D®g", role: "three-brothers" },
+    { name: "D†g", role: "three-brothers" },
+    { name: "Dπg", role: "three-brothers" },
 ];
 let token, game;
 
@@ -548,6 +551,43 @@ describe("B - Full game of 10 players with all roles", () => {
             .post(`/games/${game._id}/play`)
             .set({ Authorization: `Bearer ${token}` })
             .send({ source: "two-sisters", action: "meet-each-other" })
+            .end((err, res) => {
+                game = res.body;
+                expect(res).to.have.status(200);
+                done();
+            });
+    });
+    it("🎲 Game is waiting for 'two-sisters' to 'meet-each-other'", done => {
+        expect(game.waiting[0]).to.deep.equals({ for: "three-brothers", to: "meet-each-other" });
+        done();
+    });
+    it("👨‍👨‍👦 The three brothers can't meet each other if play's source is not 'three-brothers' (POST /games/:id/play)", done => {
+        chai.request(app)
+            .post(`/games/${game._id}/play`)
+            .set({ Authorization: `Bearer ${token}` })
+            .send({ source: "all", action: "meet-each-other" })
+            .end((err, res) => {
+                expect(res).to.have.status(400);
+                expect(res.body.type).to.equals("BAD_PLAY_SOURCE");
+                done();
+            });
+    });
+    it("👨‍👨‍👦 The three brothers can't meet each other if play's action is not 'meet-each-other' (POST /games/:id/play)", done => {
+        chai.request(app)
+            .post(`/games/${game._id}/play`)
+            .set({ Authorization: `Bearer ${token}` })
+            .send({ source: "three-brothers", action: "delegate" })
+            .end((err, res) => {
+                expect(res).to.have.status(400);
+                expect(res.body.type).to.equals("BAD_PLAY_ACTION");
+                done();
+            });
+    });
+    it("👨‍👨‍👦 The three brothers meet each other (POST /games/:id/play)", done => {
+        chai.request(app)
+            .post(`/games/${game._id}/play`)
+            .set({ Authorization: `Bearer ${token}` })
+            .send({ source: "three-brothers", action: "meet-each-other" })
             .end((err, res) => {
                 game = res.body;
                 expect(res).to.have.status(200);
@@ -1452,6 +1492,17 @@ describe("B - Full game of 10 players with all roles", () => {
             .post(`/games/${game._id}/play`)
             .set({ Authorization: `Bearer ${token}` })
             .send({ source: "two-sisters", action: "meet-each-other" })
+            .end((err, res) => {
+                game = res.body;
+                expect(res).to.have.status(200);
+                done();
+            });
+    });
+    it("👨‍👨‍👦 The three brothers meet each other (POST /games/:id/play)", done => {
+        chai.request(app)
+            .post(`/games/${game._id}/play`)
+            .set({ Authorization: `Bearer ${token}` })
+            .send({ source: "three-brothers", action: "meet-each-other" })
             .end((err, res) => {
                 game = res.body;
                 expect(res).to.have.status(200);